feat(V3): PercentageCircle: Renamed folder to V3. #434
