+++ src/components/V3/PercentageCircleV3/EthFactrIcon.tsx
